feat(swc/core): expose nodejs_common
