+++ crates/binding_core_node/src/parse.rs
@@ -14,8 +14,8 @@ use swc_core::{
         Compiler,
     },
     common::{comments::Comments, FileName},
+    node::{deserialize_json, get_deserialized, MapErr},
 };
-use swc_nodejs_common::{deserialize_json, get_deserialized, MapErr};
 
 use crate::{get_compiler, util::try_with};
 
@@ -126,7 +126,7 @@ pub fn parse(
     filename: Option<String>,
     signal: Option<AbortSignal>,
 ) -> AsyncTask<ParseTask> {
-    swc_nodejs_common::init_default_trace_subscriber();
+    crate::util::init_default_trace_subscriber();
 
     let c = get_compiler();
     let options = String::from_utf8_lossy(options.as_ref()).to_string();
@@ -149,7 +149,7 @@ pub fn parse(
 
 #[napi]
 pub fn parse_sync(src: String, opts: Buffer, filename: Option<String>) -> napi::Result<String> {
-    swc_nodejs_common::init_default_trace_subscriber();
+    crate::util::init_default_trace_subscriber();
     let c = get_compiler();
 
     let options: ParseOptions = get_deserialized(&opts)?;
@@ -186,7 +186,7 @@ pub fn parse_sync(src: String, opts: Buffer, filename: Option<String>) -> napi::
 
 #[napi]
 pub fn parse_file_sync(path: String, opts: Buffer) -> napi::Result<String> {
-    swc_nodejs_common::init_default_trace_subscriber();
+    crate::util::init_default_trace_subscriber();
     let c = get_compiler();
     let options: ParseOptions = get_deserialized(&opts)?;
 
@@ -223,7 +223,7 @@ pub fn parse_file(
     options: Buffer,
     signal: Option<AbortSignal>,
 ) -> AsyncTask<ParseFileTask> {
-    swc_nodejs_common::init_default_trace_subscriber();
+    crate::util::init_default_trace_subscriber();
 
     let c = get_compiler();
     let path = PathBuf::from(&path);
